#100 - Refactor router names, added :subject in the endpoint commented and update docker-compose for image volume
