+++ docker-compose.yml
@@ -57,6 +57,8 @@ services:
       - "8004:8004"
     networks:
       - mynetwork
+    volumes:
+      - gameservice_images:/usr/src/gameservice/public/images
     environment:
       MONGODB_URI: mongodb://mongodb:27017/game
 
@@ -130,7 +132,8 @@ services:
 volumes:
   mongodb_data:
   prometheus_data: 
-  grafana_data: 
+  grafana_data:
+  gameservice_images:
 
 networks:
   mynetwork:
